# swagger.yaml
openapi: 3.0.0
info:
  title: Pacific Telecom
  description: Apis para Pacific Telecom
  version: 0.1
servers:
  # - url: http://localhost:3000
  - url: http://138.197.105.129:3000
    description: Pacific Telecom Resource Server

# 1) Define the security scheme type (HTTP bearer)
components:
  securitySchemes:
    bearerAuth:            # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT    # optional, arbitrary value for documentation purposes
# 2) Apply the security globally to all operations
security:
  - bearerAuth: []         # use the same name as above
paths:
...
  /services/{id}:
    get:
      summary: Obtener un servicio por id
      tags:
        - Servicios
      parameters:
        - name: id
          in: path
          required: true
          description: id del servicio
          schema:
            type : integer
            format: int64
            minimum: 1
      responses: 
        '200':
          description: OK

    /services/{id}:
    delete:
      summary: Borra un servicio por id
      tags:
        - Servicios
      parameters:
        - name: id
          in: path
          required: true
          description: id del servicio
          schema:
            type : integer
            format: int64
            minimum: 1
      responses: 
        '200':
          description: OK

      /services/{id}:
    patch: 
      summary: Actualizar un servicio por id
      tags:
        - Servicios
      parameters:
        - name: id
          in: path
          required: true
          description: id del servicio
          schema:
            type : integer
            format: int64
            minimum: 1
      requestBody:   
        content:
          application/json:
            schema:
              type: object
            example: '{"name": "test"}'
      responses: 
        '200':
          description: OK
  /services/register:
    post: 
      summary: Registrar un servicio
      tags:
        - Servicios
      requestBody:   
        content:
          application/json:
            schema:
              type: object
            example: '{"name": "test"}'
      responses: 
        '201':
          description: OK


 ############################################### [ REPORTES DE FALLAS ] ############################################
...
  /server_email_config/{id}:
    get:
      summary: Obtener un reporte de falla por id
      tags:
        - Configuracion del correo del servidor
      parameters:
        - name: id
          in: path
          required: true
          description: id de la configuracion del correo del servidor
          schema:
            type : integer
            format: int64
            minimum: 1
      responses: 
        '200':
          description: OK

      /server_email_config/{id}:
    patch: 
      summary: Actualizar un reporte de falla por id
      tags:
        - Configuracion del correo del servidor
      parameters:
        - name: id
          in: path
          required: true
          description: id de la configuracion del correo del servidor
          schema:
            type : integer
            format: int64
            minimum: 1
      requestBody:   
        content:
          application/json:
            schema:
              type: object
              properties:
                 uuid:
                  required: true
                  type: uuid
                 url_server:
                  required: true
                  type: string
                 port: 
                  required: true
                  type: integer
                 status:
                  required: false
                  type: boolean
            example: '{"uuid": "ec1a3c3a-7226-11eb-9439-0242ac130002","url_server": "test","port": 31231231,"status":false, "username":"test"}'
      responses: 
        '200':
          description: OK
  /server_email_config/register:
    post: 
      summary: Registrar un Reportes de falla
      tags:
        - Configuracion del correo del servidor
      requestBody:   
        content:
          application/json:
            schema:
              type: object
              properties:
                 uuid:
                  required: true
                  type: uuid
                 url_server:
                  required: true
                  type: string
                 port: 
                  required: true
                  type: integer
                 status:
                  required: false
                  type: boolean
            example: '{"uuid": "ec1a3c3a-7226-11eb-9439-0242ac130002","url_server": "test","port": 31231231,"status":false,  "username" : "test","password" :"test"}'
      responses: 
        '201':
          description: OK

          
 ############################################### [ DATOS DE LA COMPANIA] ############################################
...